# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: KiramoPay API
  description: |
    KiramoPay is a fintech super app for Costa Rica. This API provides banking,
    payments, crypto exchange, marketplace integration, and more.
  version: 1.0.0
  contact:
    name: KiramoPay Team
servers:
  - url: http://localhost:8080
    description: Local development
  - url: http://kiramopay-api:8080
    description: Docker/Kubernetes internal
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login with cédula and PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '423':
          $ref: '#/components/responses/Locked'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout (invalidate session)
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logged out
...
  /api/v1/users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    patch:
      tags: [Users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          $ref: '#/components/responses/Success'

  # ── Wallets ────────────────────────────────────────────────────────────
  /api/v1/wallets/me:
    get:
      tags: [Wallets]
      summary: Get current user wallet
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Wallet details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/WalletRecord' }
  /api/v1/wallets/me/balance:
    get:
      tags: [Wallets]
      summary: Get wallet balance
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Balance in centimos
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/BalanceResponse' }

  # ── Transactions ───────────────────────────────────────────────────────
  /api/v1/transactions:
    post:
      tags: [Transactions]
      summary: Create a transaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    get:
      tags: [Transactions]
      summary: List user transactions
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/transactions/{id}:
    get:
      tags: [Transactions]
      summary: Get transaction by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── SINPE ──────────────────────────────────────────────────────────────
  /api/v1/sinpe/contacts:
    get:
      tags: [SINPE]
      summary: Get SINPE contacts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Contact list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [SINPE]
      summary: Add SINPE contact
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddSinpeContactRequest'
      responses:
        '201':
          description: Contact added
  /api/v1/sinpe/send:
    post:
      tags: [SINPE]
      summary: Send SINPE transfer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SinpeSendRequest'
      responses:
        '200':
          description: Transfer completed
  /api/v1/sinpe/history:
    get:
      tags: [SINPE]
      summary: Get SINPE transfer history
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Transfer history

  # ── Services & Payments ────────────────────────────────────────────────
  /api/v1/services/pay-bill:
    post:
      tags: [Services]
      summary: Pay a bill
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayBillRequest'
      responses:
        '200':
          description: Bill paid
  /api/v1/services/recharge:
    post:
      tags: [Services]
      summary: Mobile phone recharge
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RechargeRequest'
      responses:
        '200':
          description: Recharge completed
  /api/v1/services/saved:
    get:
      tags: [Services]
      summary: Get saved services
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Saved services list
    post:
      tags: [Services]
      summary: Add a saved service
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Service saved
  /api/v1/services/history:
    get:
      tags: [Services]
      summary: Get payment history
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Payment history

  # ── Crypto ─────────────────────────────────────────────────────────────
...
  /api/v1/crypto/assets:
    get:
      tags: [Crypto]
      summary: Get user crypto portfolio
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User's crypto assets
  /api/v1/crypto/transactions:
    get:
      tags: [Crypto]
      summary: Get crypto transaction history
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Crypto transactions
  /api/v1/crypto/buy:
    post:
      tags: [Crypto]
      summary: Buy cryptocurrency
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CryptoBuySellRequest'
      responses:
        '200':
          description: Purchase completed
  /api/v1/crypto/sell:
    post:
      tags: [Crypto]
      summary: Sell cryptocurrency
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CryptoBuySellRequest'
      responses:
        '200':
          description: Sale completed
  /api/v1/crypto/convert:
    post:
      tags: [Crypto]
      summary: Convert between cryptocurrencies
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Conversion completed
  /api/v1/crypto/staking:
    get:
      tags: [Crypto]
      summary: Get staking positions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Active staking positions
    post:
      tags: [Crypto]
      summary: Stake cryptocurrency
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Staking initiated
  /api/v1/crypto/staking/{id}:
    delete:
      tags: [Crypto]
      summary: Unstake cryptocurrency
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Unstaking initiated
  /api/v1/crypto/alerts:
    get:
      tags: [Crypto]
      summary: Get price alerts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Price alert list
    post:
      tags: [Crypto]
      summary: Create price alert
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Alert created
  /api/v1/crypto/alerts/{id}:
    delete:
      tags: [Crypto]
      summary: Delete price alert
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: Alert removed

  # ── Marketplace (Phase 5) ──────────────────────────────────────────────
  /api/v1/marketplace/partners:
    get:
      tags: [Marketplace]
      summary: Get available marketplace partners
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of partners (Uber, DiDi, Rappi, etc.)
  /api/v1/marketplace/connect:
    post:
      tags: [Marketplace]
      summary: Connect a marketplace partner
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [partner_code]
              properties:
                partner_code:
                  type: string
                  example: uber
      responses:
        '201':
          description: Partner connected
  /api/v1/marketplace/connect/{code}:
    delete:
      tags: [Marketplace]
      summary: Disconnect a marketplace partner
      security:
        - bearerAuth: []
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Partner disconnected
  /api/v1/marketplace/rides:
    post:
      tags: [Marketplace]
      summary: Create a ride request
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Ride created
    get:
      tags: [Marketplace]
      summary: List user ride requests
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Ride list
  /api/v1/marketplace/rides/{id}:
    get:
      tags: [Marketplace]
      summary: Get ride details
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Ride details
    patch:
      tags: [Marketplace]
      summary: Update ride status
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Status updated
  /api/v1/marketplace/food-orders:
    post:
      tags: [Marketplace]
      summary: Create food order
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Order created
    get:
      tags: [Marketplace]
      summary: List food orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Order list
  /api/v1/marketplace/food-orders/{id}:
    get:
      tags: [Marketplace]
      summary: Get food order details
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Order details
    patch:
      tags: [Marketplace]
      summary: Update food order status
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Status updated

  # ── Loyalty & Rewards (Phase 5) ────────────────────────────────────────
  /api/v1/loyalty/account:
    get:
      tags: [Loyalty]
      summary: Get loyalty account (points, tier)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Loyalty account details with tier info
  /api/v1/loyalty/transactions:
    get:
      tags: [Loyalty]
      summary: Get loyalty point transactions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Points earn/spend history
  /api/v1/loyalty/earn:
    post:
      tags: [Loyalty]
      summary: Earn loyalty points
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Points earned
  /api/v1/loyalty/rewards:
    get:
      tags: [Loyalty]
      summary: Get available rewards catalog
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Rewards catalog
  /api/v1/loyalty/redeem:
    post:
      tags: [Loyalty]
      summary: Redeem a reward
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Reward redeemed
  /api/v1/loyalty/redemptions:
    get:
      tags: [Loyalty]
      summary: Get redemption history
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Redemption history
  /api/v1/loyalty/cashback-rules:
    get:
      tags: [Loyalty]
      summary: Get cashback rules per category
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cashback percentage rules

  # ── QR Payments (Phase 5) ──────────────────────────────────────────────
  /api/v1/qr/merchant:
    post:
      tags: [QR Payments]
      summary: Register as QR merchant
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Merchant registered
    get:
      tags: [QR Payments]
      summary: Get merchant details
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Merchant info
  /api/v1/qr/codes:
    post:
      tags: [QR Payments]
      summary: Create a QR code
      security:
        - bearerAuth: []
      responses:
        '201':
          description: QR code created
    get:
      tags: [QR Payments]
      summary: Get user's QR codes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: QR code list
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
...
  schemas:
...
    RegisterRequest:
      type: object
      required: [cedula, full_name, pin, phone]
      properties:
        cedula:
          type: string
          example: "702650930"
        full_name:
          type: string
          example: "Keilor Martinez"
        pin:
          type: string
          minLength: 4
          maxLength: 6
          example: "1234"
        phone:
          type: string
          example: "+50688889999"
        email:
          type: string
          format: email
...
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            access_token:
              type: string
            refresh_token:
              type: string
            user:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                cedula:
                  type: string
                full_name:
                  type: string
...